+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+0.2.4(2019-02-26)
+------------------
+
+* Improve Config module
+* Improve rpc_call_wrap
 
 0.2.2-3 (2019-02-26)
 ------------------